refactor: Extract WebViewErrorUi to core:ui:component
