+++ core/ui/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="webview_error_description">서버 연결에 실패했습니다.</string>
+    <string name="webview_reload_text_button">다시 시도</string>
+</resources>
