+++ feature/graph/src/main/res/values/strings.xml
@@ -4,7 +4,5 @@
     <string name="graph_screen_bottom_selected_image_label">탐색 이미지</string>
     <string name="graph_screen_dialog_close_button">확인</string>
     <string name="graph_screen_dialog_title">알림</string>
-    <string name="graph_screen_network_error">서버 연결에 실패했습니다.</string>
-    <string name="graph_screen_reload_text_button">다시 시도</string>
     <string name="graph_screen_image_not_found_error">사진을 찾지 못했습니다.</string>
 </resources>
